Remove unnecessary privilege escalation in tasks
Most of the ansible.builtin.stat tasks did not require elevated
permissions because the tasks' path parameters are readable by the
ansible_user.

For the ansible.builtin.get_url tasks, the dest parameter for the files
are accessible by the ansible_user. For the task involving a GPG key,
while it may seem plausible that I would want to have this file owned by
root, this is the ascii armored key to be inserted and output in its
binary form. Meaning, I wouldn't intend on keeping it around.

The command 'netplan get all' does not require elevated permissions but
I am fine with leaving the elevated permissions in the
'Apply netplan configurations' task because the command
'netplan generate' does require elevated permissions.