Have shell tasks fail on any non-zero status
In the majority of cases, I do not want shell scripts to continue in the
event any of its commands return a non-zero status. This commit has a
good example as to why this is.

Previously, in the 'Install Kubernetes's GPG key' task, the gpg command
will fail to write the dearmor to its output file path if the file
already exists. Despite this, since the last command returns a zero
status, this causes the task itself to report 'ok' to Ansible and
proceed onward. This would be a problem later on if the installed GPG
key needed to be replaced.

The gpg command now uses --yes as an option to force overwriting of the
output file path.