Rearrange order of vars and loop keys
The vars key should be the last key within a play. This key should also
be at the bottom of tasks, proceeded at least by loop and when keys. The
loop key should also be at the bottom of tasks only to be proceeded at
least by the when key. This is a preference decision and not one of
correctness.