Migrate k8s-node packer images to Debian 12
The boot_command has changed because, for this Debian ISO, the
boot_command represents a kernel boot command and not a GRUB boot
command used by the previous Ubuntu ISO. Meaning commands like 'linux',
'initrd', and 'boot' are not valid in the former context. I will not
quote parameter values for either command types going forward unless
absolutely required because for example, quoting the value to 'initrd'
caused the kernel to be unable to find the initrd.

The 'cloud-init status --wait' command has been removed because this
Debian ISO does not utilize cloud-init. The apt-get commands have been
split because I'd prefer to split them unless I have specific reason to
conjoin them with a && control operator (e.g. Dockerfile layer
efficiency).

Any configuration differences not explained between
autoinstall.pkrtpl.hcl and preseed.pkrtpl.hcl may be attributed to
attempting to keep the enabled defaults (those not commented out) from
the example Debian bookworm (12) preseed. The locale is just en_US in
the preseed because it seems to translate to en_US.UTF-8 on the
installed system. I've discarded using ansible-client as the hostname
because this hostname should be temporary, to be replaced when used by
the deployed VM.