Rename the libvirt system user as libvirt-ro
This was done because I desired to set the group parameter in the
'Create the libvirt system user' task to that of the named user. I
discovered this would mean that the libvirt system user would have
access to the libvirt read-write socket. Which isn't desired for this
system user. Hence, I've renamed this user to be able to set group
parameter while retaining just read-only socket access.

The 'Create the libvirt system user' task shell parameter has been
changed to bash because any interactive user should use bash as their
user shell. The system parameter has been removed because since this
user is meant for interactive usage, it should not be an operating
system "system user". Moreover, the '(debian-like)' suffix has been
removed because I am not under the impression there is anything in the
task that is specific to Debian distributions.

The variables to denote the libvirt_system_user and
libvirt_system_user_home have been removed because the majority of the
ansible.builtin.user tasks do not use variables for their similar
sequence of adjacent tasks. These variables also felt pointless because
any change in their values meant the variable names would more than
likely need to be changed too.