Shed system user verbiage for non-OS system users
This was done because the 'system user' verbiage has meant to mean a
user account on an operating system. This hasn't considered that
operating systems can have actual system-level user accounts. Hence,
going forward, this verbiage will only be used if the account created is
an operating system system-level user account.