Add chk_vagrant_pkg
This script helps enforce the vagrant package used for the repository.
This was done primarily out of concern that the changes in the embedded
Ruby interpreter shipped with the package could break any custom logic
above the 'Vagrant.configure(...) do' block.

.shellcheckrc was added to allow the following of arbitrary file paths
because this cannot be enabled in individual scripts.

The constant VAGRANT_UPSTREAM_VERSION name was inspired by the
'upstream_version' that is part of the 'Version' field mentioned in the
debian-policy. The constant value will exclude the '-debian_revision'
suffix because said suffix is specific to deb packages.