Fix virtual NICs to use DHCP in packer images
ssh_password is now ansible_user_password because packer never used
password authentication when provisioning the targeted machines. Hence
the former variable name/prefix/suffix was misleading.

The shell provisioner has been replaced with the ansible provisioner
because the provisioning logic is now more complex than just upgrading
operating system packages. The use_proxy parameter is set to false
because by default this is true, causing packer to be unable to connect
to the machines to provision them. The inventory_file_template parameter
follows mostly what is mentioned as default in its parameter
description.

config_refs that are an Array or Hash will now be converted to JSON and
quoted to be digestible by Ansible, as these do become host variables.

To keep consistent with production, the vagrant/staging machines will
also now utilize systemd-networkd. Because the network interface cards
(NICs) used for the poseidon-k8s-homelab-cm network in staging are set
with a static IP address, a general glob pattern could not be used for
the NICs like in 'k8s-nodes.pkr.hcl'.

ansible user within the 'Append the ansible user to the libvirt group'
task is now ansible_user because I am referring to whoever the
ansible_user is set as (e.g. vagrant, ansible) and not a literal
'ansible' user.