Add the maintenance.yml playbook
The jmespath Python dependency has been added because it's required for
the community.general.json_query filter. The psutil Python dependency
will now be added to the k8s_controllers and k8s_workers client Python
virtual environments because this is a requirement for the
community.general.pids task module.

Settings used throughout the backend availability plays derive from the
respective HAProxy configurations (e.g. timeouts, ca-file). The
ansible_hostname is used as part of the url parameter to the 'Check
connection to the /readyz endpoint ...' task because the kube-apiserver
TLS certificate generated by Kubernetes does not contain 127.0.0.1 or
localhost as a subject alternative name. Whereas, the hostname of the
controller the kube-apiserver is running on is one of the subject
alternative names.

A lot of the "whys" for the 'Upgrade Kubernetes cluster (poseidon)' play
can be explained by the Kubernetes documentation when upgrading kudeadm
clusters
https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/kubeadm-upgrade/.
Calico has been upgraded according to
https://docs.tigera.io/calico/3.24/operations/upgrading/kubernetes-upgrade#all-other-upgrades.

The apt-daily and apt-daily-upgrade systemd units (e.g. timer, service)
are stopped, disabled, and then reenabled because this process itself or
child processes can hold the dpkg lock for a notable amount of time.
Through testing, I found this can lead to any of the tasks utilizing
apt or dpkg to fail. Without intervening on these units, it was more
than likely to expect that the maintenance.yml playbook would fail at
least once, if not more. SIGTERM is sent to the apt-daily-upgrade
service process group because KillMode is set within the service to
'process', which means upon stopping the service, only the main process
itself is killed (child processes remain).

Child processes apt.systemd.daily and unattended-upgrade of the
apt-daily-upgrade service are explicitly killed in addition because
through testing, I've found that the mentioned service can spin up
between the 'Send SIGTERM to apt-daily-upgrade process group' and 'Stop
and disable the apt-daily-upgrade service' tasks. This assumes that the
apt-daily-upgrade service was not active at time of the 'Get process ID
of the apt-daily-upgrade service' task. These processes still being
active causes the same problem as previously mentioned.

The apt-daily and apt-daily-upgrade systemd services are not started
towards the end of the 'Upgrade Kubernetes cluster (poseidon)' play
because these services are meant to be triggered on their respective
systemd timers. Despite the state not differing, it is still added as a
parameter to keep consistent with the rest of my Ansible code.

The 'Mark the Kubernetes nodes as schedulable' task is placed under the
'always' block to prevent any failing tasks from causing the cluster to
become defunct due to nodes being unschedulable.

The kubeconfig files are copied again because certificate information
changes in these files upon upgrading.

The Poseidon Kubernetes software versions have been incremented to test
the new mentioned playbook out. Otherwise, 'kubeadm upgrade apply' would
have error out due to having an issue with the kubeadm version not
matching the Kubernetes version for the cluster. The version mismatch
was intentional due to what was seen in 264e7357.