Add capability to limit hosts for playbooks' runs
ANSIBLE_LIMIT is not explicitly set to be exported like
ANSIBLE_EXTRA_VARS and ANSIBLE_MAINTENANCE_PLAYBOOK because I found that
Makefile variables set via command line are already exported.

ANSIBLE_PLAYBOOK_OPTIONS is a simply expanded variable because the value
to the Makefile append operator must be expanded, which does not happen
if ANSIBLE_PLAYBOOK_OPTIONS was a recursively expanded variable. If
ANSIBLE_PLAYBOOK_OPTIONS was a recursively expanded variable, when
ANSIBLE_PLAYBOOK_OPTIONS is finally expanded, it may contain additional
whitespace where additional options are added.

The production and production-maintenance rules' recipes have been
rewritten to handle optional command options better, as continuing to
use ifeq conditionals would not be scalable.