Reorganize the top-level playbooks' running order
This was done because before these changes, the playbooks' running order
coupled with the provisioning order. This coupling is now not necessary
because of the added limit capability that limits hosts for playbooks'
runs. My DHCP servers, DNS servers, load balancers and virtual balancers
and virtual machine monitors are part of my homelab. Hence, their
playbooks should be imported as part of the site.yml playbook.