Set Kubernetes cluster (Poseidon) domains state
This was done because the default state of a domain once defined is to
be off, even if the autostart parameter is set to true.

Setting the state of a domain has to be in a separate task because the
state and command parameters of the community.libvirt.virt task module
are mutually exclusive.