Use an expert_recipe for k8s-nodes packer images
This was done because the atomic recipe creates a swap space partition,
and for the kubelet to work properly, any swap space must be disabled.
Currently, there exists some code to work around this in the k8s_kubeadm
role, but since I have control of how these images are built, I believe
it would be better not to create a swap space partition.

The 'd-i partman-basicfilesystems/no_swap' configuration has been added
to automate an answer to a new question warning about how there will be
no configured swap space partitions.