Refactor Renovate configurations
The $schema attribute has been added because this has been added to a
onboarding configuration for Renovate on GitHub.

config:base has been changed to config:recommended because the former
has been renamed to the latter upstream. The PR for this would be here,
https://github.com/renovatebot/renovate/pull/21136.

The timezone attribute has been removed because this isn't something I
should think about if there are no schedules defined.