Add CronJobs to the Kubernetes cluster (Poseidon)
The 'cronjobs' namespace is hardcoded because I do not feel using a
variable is currently necessary.

The 'type' field is omitted from the Secret object created as part of
cronjobs.yml.j2 because Opaque is the default Secret type and is the
Secret type I intend to use for said object. This is despite what I've
done in the past for Secret objects, such as those integrated in
ff8c695.