Migrate to just ansible-core and collections used
This was done to optimize what's installed for Ansible, by installing
only what's used. In theory, a collection update means only the Ansible
units using that collection need to be tested, this can lead to quicker
testing.

collections_path is set within ansible.cfg to localize the collection
dependencies to this repository and to prevent a different version than
expected of a collection from being used.