Remove global like directories from roles_path
This was done because any role invoked should be defined or installed
separately within the repository. The same rationale is why the
ANSIBLE_ROLES_PATH environment variable changed within 'molecule.yml'.

The trailing colon has been removed because in Ansible, a trailing or
leading colon can result in a path that includes the base directory of
the ansible.cfg file. In this case, that is not necessary.