Utilize Ansible's DEFAULT_LOG_PATH for logging
This was done to let Ansible handle the logging of output locally,
replacing the makeshift solution I had using shell redirection. These
changes simplify the Makefile and upgrades the logging done, as the
ansible-playbook runs that are part of building the packer images can
now have their output logged. This wasn't previously possible (at least
not without a lot of duct tape) because simply redirecting packer's
stdout/stderr would have a mix output of both packer and the
ansible-playbook runs.

A timezone_offset variable had to be added to k8s-nodes.pkr.hcl because
the timestamp function used by packer returns a timestamp string that
follows the RFC 3339 syntax. Timestamps following the RFC 3339 syntax
are in the UTC timezone. This variable is used with the timeadd function
to convert the timestamp into one that is more closely locale-aware.

A unique suffix is appended to each log file name, primarily because I
want each playbook run to generate a separate log file. This becomes
problematic in the case of the packer images, as they are provisioned
roughly at the same time. Meaning, it would be possible that each image
built could use the same log file. Such a suffix isn't necessary for
other Makefile targets that do ansible-playbook runs, but for
consistency, I do want them to also have the same kind of suffix.