Use ANSIBLE_VERBOSITY over ANSIBLE_VERBOSITY_OPT
This was done because setting ANSIBLE_VERBOSITY from the command line
can provide the same effect as using ANSIBLE_VERBOSITY_OPT, the latter
requiring custom integrations. The former will work by default with all
invocations of ansible that are through the Makefile rules, packer, and
vagrant.

ANSIBLE_VERBOSITY is set to 0 for the ansible command within the
k8s-node-images rule because I don't typically want verbose output for
said command, even if ANSIBLE_VERBOSITY is set from the command line.