Do not set a transient hostname from DHCP servers
This was done because systemd-networkd will, by default, attempt to set
a hostname received by a DHCP server as the transient hostname for a
host. Letting systemd-networkd do this wouldn't be an issue, but
currently, this generates a 'Could not set hostname: Access denied'
error when examining systemd-journald logs. Hence, I find letting this
error continue to be emitted would be leaving in unnecessary noise when
shifting through logs.