Set infinite lease for poseidon-k8s-homelab-cm hosts
This was done because with libvirt using dnsmasq, the default DHCP lease
time is one hour. Letting DHCP leases expire every hour for hosts within
the poseidon-k8s-homelab-cm network causes the Kubernetes cluster
(Poseidon) keepalived instances to have unnecessary elections. When a
host's lease expires, keepalived will unassign that DHCP reserved IP
address from the interface that also has keepalived's virtual IP (VIP)
set to it. This causes an instance to enter the FAULT state, possibly
removing the VIP set to it and forcing the other instances to hold an
election for the next master instance. A pattern of elections like this
introduces instability and and has led to confusion when troubleshooting
other components.