Set k8s cluster nodes NICs' model type to virtio
This was done because it was noticed in production that the "default"
model type (rtl8139) used lead to significant instability for the
Kubernetes cluster (Poseidon), as the cluster network speed was
insufficient. I've made this change to all network interface cards
(NICs) as there isn't a reason to use the default model type. To keep
consistent with production, the staging VM templates will also have
this change (vagrant-libvirt sets domains' model type to virtio by
default).