Migrate k8s cluster (Poseidon) to use cgroup v2
A containerd configuration file is created (instead of being removed)
because it is recommended in the official Kubernetes documentation that
containerd uses the systemd cgroup driver to set cgroup parameters. This
configuration file is generated dynamically instead of coming a Jinja
template because it was either this or maintain a template that contains
"opinionated" defaults with my single change.