Use a local hook for the prettier pre-commit hook
This was done because the pre-commit/mirrors-prettier repository has
been archived, with no official replacement taking its place.
Furthermore, with this, prettier has removed this pre-commit option has
part of their documentation. The discussion I followed for these events
can be found here, https://github.com/prettier/prettier/pull/16229.

The additional_dependencies for prettier are not formatted like those
for ansible-lint because these packages are eventually passed to npm to
be installed. Hence, they must be formatted as a npm package-spec
argument. The version of prettier appended was derived from
'package-lock.json'.