Rename shell scripts to match my personal style
Shell scripts that are meant to be executable will use dashes and
will continue not using a file extension. This was done because I don't
typically associate commands containing an underscore as an executable.
Instead, I'd be more likely to think a command containing an underscore
to be a shell function.

Shell scripts to be read in/sourced will not contain a shabang line, and
have a file extension of '.sh'. This was inspired by how Google treats
libraries in their shell style guide, which can be found here,
https://google.github.io/styleguide/shellguide.html#file-extensions.
The disable-make-resolv-conf.sh script is an exception to this on a host
because another executable used by dhclient (run-parts) does not accept
files named with '.' characters (see run-parts(8)).

The disable-make-resolv-conf.sh script had its shabang line removed
because this file is sourced by dhclient and not actually executed.