Modify approve CSRs task changed_when, failed_when
The changed_when conditional has been modified to use the length Jinja
filter as this is what's used elsewhere in the codebase when having a
boolean expression involving a non-empty or empty string.

The failed_when conditional has been removed because I want this task to
fail when the command exits with a non-zero exit code. Conveniently,
this is already how failed_when is defined by default for the
ansible.builtin.command task module. I derived this through trial and
error, plus ansible's source code.