Move installing jq deb pkg closer to where needed
This was done for the k8s_post_controllers.yml and maintenance.yml
playbooks because without having both these playbooks install the jq deb
package, coupling would exist between both playbooks. I believe coupling
exists, as removing the Metrics Server would most likely also lead me to
remove the task installing jq (see b276c4d0).

This was done for the systemd_networkd role because this should be a
package installed by the role itself, not relying on the on_prem.yml
playbook to install it.

The comment from the 'Install controller-wide tooling (debian-like)'
task has not been retained because now that context is provided for why
the jq deb package is installed, it makes the comment unnecessary.