Start automating conventions with existing tooling
The conventions 'Single quote regular expression or glob patterns.' and
'Double quote octal numbers to the mode parameter.' will be merged in
favor of enforcing just using double or no quotes with yamllint. This
means most of the regex and glob patterns lost their quotes or had them
replaced with double quotes. I will be entrusting that yamllint can
determine the quote style (e.g. single, double, none) to use for
functional Ansible yaml. I favored this over trying to automate
both conventions because this would require a more custom solution that
I didn't feel it would be worth the effort.

The convention 'Be direct when possible when referring to operating
system users in task names.' has been rewritten to 'Hardcode the
operating system user's name in the task name.' as I found the former
to be vague and wasn't sure initially what I intended to enforce with
this convention.

In the maintenance.yml playbook, many 'when' conditionals were
double-quoted, but through testing I found the quoting to be
unnecessary. The commit where the quoting was introduced, 8326914,
doesn't provide an explanation for the quoting. Hence, I'm removing it.

ansible-lint is added as an additional dependency for mypy's pre-commit
hook because there are types I use from ansible-lint's codebase.

Additional Ruff rules have been added via the extend-select setting
because I believe these rules will be useful along with the default
rules.