Start automating conventions with existing tooling
The convention for setting the errexit and pipefail options for
ansible.builtin.shell tasks has been rewritten in its rule to give
better guidance of what's expected.

The conventions 'Single quote regular expression or glob patterns.' and
'Double quote octal numbers to the mode parameter.' will be merged in
favor of enforcing just using double or no quotes with yamllint. This
means most of the regex and glob patterns lost their quotes or had them
replaced with double quotes. I will be entrusting that yamllint can
determine the quote style (e.g. single, double, none) to use for
functional Ansible yaml. I favored this over trying to automate both
conventions because this would require a more custom solution that I
didn't feel it would be worth the effort. Utilizing yamllint in this
manner is also why the convention 'In general, do not use quotes in the
ansible yaml.' has been deleted.

The convention 'Be direct when possible when referring to operating
system users in task names.' has been rewritten to 'Hardcode the
operating system user's name in the task name.' as I found the former
to be vague and wasn't sure initially what I intended to enforce with
this convention.

The convention 'Follow the indentation used through the codebase for all
`ansible` yaml.' will be enforced by the default opinionated indentation
of ansible-lint.

Despite what
https://github.com/ansible/ansible-lint/pull/1144/commits/673617d7
mentions, I found the only way to load the ansible-lint testing fixtures
was to give pytest_plugins an import path of
'ansiblelint.testing.fixtures'.

In the maintenance.yml playbook, many 'when' conditionals were
double-quoted, but through testing I found the quoting to be
unnecessary. The commit where the quoting was introduced, 8326914,
doesn't provide an explanation for the quoting. Hence, I'm removing it.

ansible-lint is added as an additional dependency for mypy's pre-commit
hook because there are types I use from ansible-lint's codebase.

Additional Ruff rules have been added via the extend-select setting
because I believe these rules will be useful along with the default
rules.

The URL referenced in the custom-no-changed-when rule has been shortened
from its original form in the documentation to better fit the error
message pop-up.

The print_modules.py script is used to assist me in determining the task
modules that require distribution specificness.

Additional entries have been added to the exclude_paths setting within
the .ansible-lint file because otherwise the filenames rule would have
thrown violations for these entries.