Create a custom key-order rule
This was done because the built-in key-order rule has almost all the
functionality I want, but only uses a single variable to define the
order of keywords for both plays and tasks. Hence, I've had to override
the KeyOrderRule class's matchplay and matchtask methods with rewritten
functions to patch this variable accordingly.