Add ACPI for the k8s cluster nodes
This was done because it was noticed in production that the nodes could
not be shutdown with 'virsh shutdown' and instead required a variant of
'virsh destroy'. I found that ACPI is required if I wish todo a graceful
shutdown of each node. To keep consistent with production, the
staging VM templates will also have this change (vagrant-libvirt already
adds ACPI to domains by default).