Rename the functions' variable _id to id_
This was done as the variable is just used a key to fetch the
appropriate violation message of a rule and doesn't have a stronger
association (e.g. _id is used as an iterator variable for _ids) with
each rule's _ids class variable. Also, I already use the variable name
id_ elsewhere in the codebase as 'id' is a Python keyword with PEP8
recommending a trailing underscore to avoid such conflict.