Integrate kiwix-serve into k8s cluster (Poseidon)
A usage attribute has been added to the nfs_exports_config group
variable because I did not want the querying of the NFS export to use to
be based on the path it's mounted to. Doing so wouldn't make a lot of
sense. Also, said variable has been moved to the poseidon group because
the poseidon_k8s_controllers group now needs access to this variable to
apply the kiwix-serve manifests.