Use ingress-nginx namespace task registered var
This was done over using an entirely separate variable for the namespace
because as of lately, I've preferred using a registered variable of an
associated task related to an attribute (e.g. namespace). In addition,
the registered variable name is now generic, so any changes to the
namespace will not affect the registered variable name.