Create convention to use files dir for copy module
This was done to utilize the local relative path searching when
specifying the src parameter to the ansible.builtin.copy task module.
Any files used by the ansible.builtin.copy task module were also moved
under this directory to declutter the top-level playbooks directory.
Doing this aids in following Ansible's documented collection structure,
even if the top-level playbooks directory is not part of a collection.