Use CNAME vs A DNS records for services on hosts
This was done because a common use case for CNAME records is for
subdomains that point to the same server. gerald and staging-node1 being
the actual hosts/servers with an A record. To add, the ping executable
will now output the actual host if an alias is pinged instead. As
previously, for example, pinging kiwix.homelab.cavcrosby.net would
output proxy.homelab.cavcrosby.net instead of
'gerald.homelab.cavcrosby.net'.