Orchestrate DDNS functionality from k8s cluster
The nfs-homelab.exports.j2 Jinja template has had its Jinja comment
removed, so the exported devices are not glued together all on one line.
I presumed I never tested more than one device being exported with this
comment in place. An extra newline is still required, though.

The --split-exp option to the yq container is quoted the way it is
because this option is also treated as an expression to be evaluated by
yq. The --output-format option was specified as yq doesn't recognize the
'txt' file format extension. This results in yq assigning the input and
output to be yaml, which isn't desired.

The default LVM logical volume size is its volume group's physical
extent because since the cache logical volume, as of this commit, holds
less than one megabyte of data, I felt any other value used would be
arbitrary. To add, logical volumes cannot be smaller than the volume
group's physical extent.

The boto3 package is installed with the AWS Common Runtime (CRT) extra
because of the touted performance and minimal footprint benefits.