Add shellcheck's pre-commit hook to the repository
A source shellcheck directive has been added to pass the shellcheck
pre-commit hook. Previously, no directive was written for the
/etc/os-release and /usr/lib/os-release paths because they passed under
the shellcheck VSCode extension. This is expected because shellcheck
uses Docker to run its pre-commit hook. For now, it is intended that I
will solely rely on shellcheck's pre-commit hook. This is because
setting a shellcheck executable path for the extension is quite
cumbersome in a VSCode multi-root workspace.