Integrate vpn_servers into the infrastructure
A localhost inventory file and playbook have been created and are run
separately together primarily because otherwise, any play using the
'all' hosts pattern would also be run on the localhost. This wouldn't be
intended as I currently associate the localhost with my development
machines. The localhost.yml playbook deviates from other playbooks in
regard to installing required operating system packages (e.g. systemd
package for systemd-networkd service and the libvirt-dev package for the
libvirt-python Python package) on the target host. In that, I purposely
don't in the localhost.yml playbook, as I don't believe it is in the
purview of this repository to manage packages installed on my
development machines.

The vpn_network_subnet is different for the production and staging
environments to be able to test the VPN using the public DNS record
assigned to it for each environment. In staging's case, to be able to
resolve its public VPN DNS record and use the staging homelab network as
its vpn_network_subnet, I would have shut off its associated virtual
network bridge on my development machine. This would be problematic
because this results in none of the VMs on that network being able to
communicate with one another. Hence, a new subnet was created that my
development machine wouldn't have access to, but the VMs would.

A script to configure iptables rules on my development machine to route
specific traffic to my staging VPN servers is now required. This is
because when traffic circles back from my networking equipment to my
development machine, the destination IP address will be for it. Whereas
that specific traffic should be rerouted to the staging VPN servers.

dnsmasq now listens on a staging VPN IP address to be able to test
querying for the staging DNS records while connected to the VPN. This is
especially true when using staging's public VPN DNS record. As this
might involve toggling off a network interface device's DNS
configuration to force my development machine to use the public VPN DNS
record.

The Kubernetes cluster (Poseidon) nodes' 'cpu' modes has been set to
host-passthrough to get around a glibc error of 'CPU does not support
x86-64-v2' emitted from an aws-cli container on a poseidon_k8s_workers
host.

The systemd_networkd role's systemd.network.j2 template has been renamed
to reflect that this template is used by other types of systemd-related
programs aside from systemd-networkd. This template is linked from
within the ./playbooks/templates directory, as it's a convenient and
easy way to reuse this template within my playbooks. Finally, this
template was rewritten to support multiple sections being repeated
within a configuration file and with an emphasis on programmatic
readability (e.g. no weird spacing within Jinja expressions).