Update Computer Graphics StackExchange ZIM url
The --fail option is added to the curl container args because otherwise,
it was noticed that if the ZIM file targeted does not exist on the
server, then curl would output and save a 404 HTML document. Unbeknownst
to me, there would be an issue until the kiwix-serve pod starts crashing
being unable to read the affected ZIM file.