Partially replace jq processor with yq processor
I did this to show more support to yq by using it over jq, even though I
still intend on using jq on my managed nodes because there is not a
native package for yq in Debian's repositories. I ended up using yq
recently due to needing to output a JSON blob to a file without using
the shell todo this (e.g. redirection or pipe to an executable that can
write to a file). Unfortunately, jq doesn’t offer an option like this,
much like the output option for curl. Also, I went down the rabbit hole
to find that the maintainers of jq would be resistant to the addition of
this feature.

Additional miscellaneous changes were done to adhere to conventions I
follow for my shell code (e.g. stdin redirection appended at the end of
a command, use quoting where possible).