Group the dependencies' updates done by Renovate
This was done to reduce the number of dependency update commits that are
eventually merged into the default branch's commit history.

The custom regular expression package managers/managers were written to
enable Renovate to manage updates for Packer, Packer plugins, and
Vagrant. The versioning for Vagrant is set to loose because, unlike
Packer, Vagrant doesn't have an explicit setting denoting the version of
it that should be used. The versioning for Packer and its plugins is set
to 'hashicorp' because Packer's version setting is very similar to the
one used by Terraform, which is what this built-in versioning was
created around.

A package rule to disable updating packages from the
mitmproxy/requirements-dev.txt file has been created to prevent Renovate
from updating packages to a version that doesn't support the system
Python version used by the proxy_servers hosts (556b70d). Renovate can
enforce a language constraint for package updates. However, this cannot
be applied per package rule, which means I couldn't restrict just these
dependencies to a Python version constraint.