Remove type ignore comment for SpellChecker import
I did this as the type ignore comment is unnecessary for the
pyspellchecker package, and its addition was likely due to oversight.
The type ignore comment is unnecessary because the package provides
typing inline with its codebase. I believe this was an oversight, as I
would only have thought it was necessary because the mypy pre-commit
hook did not have the package as an additional dependency.