Fix tracking the OpenTofu deps as Terraform deps
This was done because, by default, Renovate will track the OpenTofu
dependencies (e.g. providers, the tofu binary) as if they were Terraform
dependencies. While there are remedies to this, currently, there is no
OpenTofu Renovate manager. The matchDatasources solution using the
OpenTofu registry is what was recommended from Renovate's documentation
for part of the OpenTofu dependencies. As for the tofu binary itself,
this dependency is handled by an already created regex custom manager.
In addition, a rule was created to prevent the binary from being picked
up as the hashicorp/terraform package too.