Fix pre-commit installing latest ansible-core pkg
I did this by pinning the version of the ansible-core package denoted in
the requirements.txt file. This was done so my Python virtual
environment and the one set up by pre-commit use the same package
version. When updating all my development dependencies, I noticed when
running the pytest executable that all my tests passed in my Python
virtual environment. This wouldn't be the case, however, in the one set
up by pre-commit.