Implement the custom-var-naming[no-reserved] rule
This rule helps enforce the usage of trailing underscores for variables'
names that conflict with Ansible keywords. As mentioned in f9e4e3d.

The custom-no-changed-when rule has been changed to alter each
MatchErrors' tag to be the rule's ID, as this prevents obscure bugs like
appending the original rule to be skipped, but it also skips the custom
rule. This wasn't an issue with this rule but was noticed at first with
the new custom rule implemented.