Use 'this' as the default resource name
This was done because the community AWS and example Talos Terraform
modules (see https://github.com/terraform-aws-modules and
https://github.com/siderolabs/terraform-provider-talos/tree/main/examples
respectively) use this keyword for resources that are the only ones of
their type. In my case, this is true and if I cannot think of a
meaningful name for a resource that is the only one of its type. I
prefer this to using 'main' as the default resource name.