Reseat allocated IP addrs to remove the addr gap
This was done primarily for the staging environment as the
address (addr) gap between its homelab network's gateway IPv4 address
and its DHCP range's lower bound for the same network came to be from
the changes in d442e4d.

As for the production environment, the 192.168.0.2 IPv4 address was not
allocated, I believe due to wanting to save the address for a layer 3
switch that would have additional physical ports for me to use. Though
as of this writing, I have no intentions to use it anytime soon.