Refactor libvirt networks' names to match domains
This was done to replace the homelab-cm suffix in the networks' names
with the libvirt prefix I use for domains. Also, since the expectation
as of this writing is that any environments' infrastructure (compute,
network, or storage) will be managed by the homelab-cm project (which
includes this repository), the prefix was not added to the network names
in my environments' infrastructure.

The management network name is no longer explicitly named in the code.md
Markdown file to be straightforward in the point being conveyed.